Add Laravel route sampling rules and slim queued job attributes
